Airflow/requirements.txt
apache-airflow-providers-airbyte==4.0.0
apache-airflow-providers-dbt-cloud==3.0.0
dbt-core
dbt-bigquery